requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyswisseph==2.10.3.2
pydantic==2.5.2
pytz==2024.1
python-dateutil==2.8.2